Jasmin John development branch
-On Sept.26, 2016 Jasmin wrote:
The latest COBALT version is in
   /home/jgj/fre/updates/COBALT/generic_COBALT_cmor_20160922.F90

This has a lot of updates for CMORized diagnostics, and should
be similar to what Bill was using, except:
 - includes a fix for Fe calculation to make sure we can reproduce
   old iron calculation if do_fan_dunne is not activated

 - does not include the c14/abiotic updates in that John D/K were
   working on in user/jpk/generic_abiotic_20160916
   These c14/abiotic updates/fixes would have to be merged in
   to the code above.

For compatibility with g_send_data/g_register_diag_field calls, I use:
  /home/jgj/fre/updates/COBALT/generic_tracer_utils_cmor_20160808.F90

The update I usually add to MOM6_generic_tracer.F90 (to set CASED
to zero below k=1), may have made it's way into the MOM6 code
recently, but if it didn't please let me know.
It's possible we do not need it, but I have not had time to run
a test to verify that the remapping does not give spurious
values below k=1 for this field.