Fixes for gs fluxes and new cmip6 diags from John Dunne
- On 2017/07/28 John Dunne wrote:
I have a revised version of the CMIP6 diagnostics for BLING in:

/lustre/f1/unswept/John.Dunne/mdt/20170519/cm4_libs_compile/src/ocean_shared/generic_tracers/generic_BLING.F90

That includes a bunch of diagnostic bug fixes that I would like to make the trunk version.  I believe the switch to generic gas fluxes changes answers, but that should be the only thing.

Note that there are still some problems with the z-remapping of four flux fields (expc, expcalc, expfe, and expp) that we will have to correct at a later date.