Fixes for interface diagnostics using MOM6 axeTi
The register_diag_field interface needs to be extended to take the MOM6 axes_grp as argument
instead of this integer array axes_grp%handle
Crently the actual MOM6 diag axes is chosen to be T or Tl based on the size of the axes argument, 2 or 3.
The actual values of these axes argument are not used, only their size is checked to determine the diag axes!
This is not correct since axesTi and axesTl are both of size 3, likewise there are many axes of size 2.
To accomodate axesTi with the least amount of code modification we can set and check for an input array of size 1.

Again, this is a hack and the register_diag_field wrapper in generic_tracer_utils has to be extended to
accomodate any MOM6 axes, by passing the string like 'axesTi'i etc rather than the handles.