Add deposition fluxes for alk
- This updates add wet and dry deposition fluxes for tracer alk and
registers diagnostics for them
- atmos_phys repo branch alkalinity_deposition_flux  atmos_dust.F90
  prepares these fluxes on the ATM side to be received by alk.
- Currently these fluxes are set to 0 till the alkalinity content of
  dust fluxes is formulated in line 956 of atmos_dust.F90
  https://gitlab.gfdl.noaa.gov/fms/atmos_phys/-/commit/7c086a6ab789ffec667b58ce220be1668348bb2f#dd6d8ddd1fbd48f1254dda4c25d144c8d2181242_919_956
- John Dunne wrote:
  "I have been talking with Charlie and Jasmin about running some ESM4.2
simulations where we add purposefully add alkalinity to the ocean induce
Carbon Dioxide Removal from the atmosphere, and we think the most
code-efficient way to do it would be to add connectivity thought the
coupler like we do for NO3 and Fe.  Would you be able to help me with
this?  The default would be to assign a zero flux over the ocean, and
the data override would be to assign a map of wet or dry deposition as
an alkalinity flux.  Jasmin and I discussed an alternative of creating a
3D data override flux within the COBALT code, but upon further thought I
believe that approach would both be much more work and less applicable
to the long term goal of being able to simulate potential alkalinization
of the atmosphere (which I think will also be proposed soon)."